Updated all classes. Removed CustomCornCost and replaced with Custom class; these classes are identical but renamed for easier readability. Removed CornProfit and replaced with Model class; identical but renamed due to eaiser readability.
Added comments to as many different methods as was needed.

Condensed a few methods for faster execution and to have less lines of code

the Custom class is still really dense because it deals with a lot of UI components that are individually handled.

Two new apk's were created with the way the code stands currently.